Delete SchoolGroupMemberships which are no longer provided in import data
This is important because, if we don't have a way to do this then:

If the old schoolgroup has MAT or LA level access, they will retain access to publish vacancies on behalf of the school even after the school is no longer part of their schoolgroup

It makes analytics calculations way more complicated than they need to be, because we have to deal with the concept that schools can be part of more than one trust or LA (for example, in a table of schools, we can no longer have a simple 'which trust is this in' column, which is useful to break things down for comms)

We delete leftover/'old' SchoolGroupMemberships separate from the import task
subclasses all at once, after running ImportSchoolData and ImportTrustData, in
order to avoid baking in the assumption that a school can only be part of a
single SchoolGroup, which may not be a valid assumption now or in the future.

As a result, we handle the deletion of old memberships by marking the memberships
which are the result of the current task with a do_not_delete attribute of 'true',
leaving the rest as false.